Refactor: Create FunctionGenerator class for function-related IL generation
- Extract DeclareFunctions and DeclareFunction methods from ILMethodGenerator to new FunctionGenerator class
- Update MainGenerator to use FunctionGenerator for function declarations
- Maintain backward compatibility and FirstMethod handling
- All 29 function tests continue to pass, confirming no regression
- Prepares architecture for nested functions implementation
